Refactor package into src layout
