+++ src/pc_proof/__init__.py
@@ -1,0 +1,1 @@
+"""Propositional logic toolkit."""
