+++ src/pc_proof/logic.lark
